Move saving template backend request to repository
